Добавлена операция подтверждения получения статусов
